backend/requirements.txt
fastapi
uvicorn
python-multipart
gitpython
radon
javalang
libclang
networkx
pydantic
